// packages/bespoke/lib/game/EyeTest/interface.d.ts
export interface ICreateParams {

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);

    /** CreateParams exchangeRate */
    exchangeRate?: (number|null);

    /** CreateParams timeLimit */
    timeLimit?: (number|null);
}
// ...
export interface IGameState {

    /** GameState gameStage */
    gameStage?: (number|null);

    /** GameState time */
    time?: (number|null);
}
// ...
export interface IPlayerState {

    /** PlayerState seatNumber */
    seatNumber?: (number|null);

    /** PlayerState anwsers */
    anwsers?: (PlayerState.IAnwser[]|null);

    /** PlayerState result */
    result?: (PlayerState.IResult|null);
}

/** Represents a PlayerState. */
export class PlayerState implements IPlayerState {

    /**
     * Constructs a new PlayerState.
     * @param [properties] Properties to set
     */
    constructor(properties?: IPlayerState);

    /** PlayerState seatNumber. */
    public seatNumber: number;

    /** PlayerState anwsers. */
    public anwsers: PlayerState.IAnwser[];

    /** PlayerState result. */
    public result?: (PlayerState.IResult|null);

    /**
     * Creates a new PlayerState instance using the specified properties.
     * @param [properties] Properties to set
     * @returns PlayerState instance
     */
    public static create(properties?: IPlayerState): PlayerState;

    /**
     * Encodes the specified PlayerState message. Does not implicitly {@link PlayerState.verify|verify} messages.
     * @param message PlayerState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IPlayerState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified PlayerState message, length delimited. Does not implicitly {@link PlayerState.verify|verify} messages.
     * @param message PlayerState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IPlayerState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a PlayerState message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns PlayerState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): PlayerState;

    /**
     * Decodes a PlayerState message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns PlayerState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): PlayerState;

    /**
     * Verifies a PlayerState message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a PlayerState message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns PlayerState
     */
    public static fromObject(object: { [k: string]: any }): PlayerState;

    /**
     * Creates a plain object from a PlayerState message. Also converts values to other types if specified.
     * @param message PlayerState
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: PlayerState, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this PlayerState to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

export namespace PlayerState {

    /** Properties of an Anwser. */
    interface IAnwser {

        /** Anwser emotion */
        emotion?: (number|null);

        /** Anwser gender */
        gender?: (number|null);

        /** Anwser time */
        time?: (number|null);
    }
// ...
    interface IResult {

        /** Result genderNum */
        genderNum?: (number|null);

        /** Result emotionNum */
        emotionNum?: (number|null);

        /** Result point */
        point?: (number|null);
    }
// ...
}

/** Properties of a MoveParams. */
export interface IMoveParams {

    /** MoveParams seatNumber */
    seatNumber?: (number|null);

    /** MoveParams newanwser */
    newanwser?: (MoveParams.INewAnwser|null);
}

/** Represents a MoveParams. */
export class MoveParams implements IMoveParams {

    /**
     * Constructs a new MoveParams.
     * @param [properties] Properties to set
     */
    constructor(properties?: IMoveParams);

    /** MoveParams seatNumber. */
    public seatNumber: number;

    /** MoveParams newanwser. */
    public newanwser?: (MoveParams.INewAnwser|null);

    /**
     * Creates a new MoveParams instance using the specified properties.
     * @param [properties] Properties to set
     * @returns MoveParams instance
     */
    public static create(properties?: IMoveParams): MoveParams;

    /**
     * Encodes the specified MoveParams message. Does not implicitly {@link MoveParams.verify|verify} messages.
     * @param message MoveParams message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IMoveParams, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified MoveParams message, length delimited. Does not implicitly {@link MoveParams.verify|verify} messages.
     * @param message MoveParams message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IMoveParams, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a MoveParams message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns MoveParams
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): MoveParams;

    /**
     * Decodes a MoveParams message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns MoveParams
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): MoveParams;

    /**
     * Verifies a MoveParams message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a MoveParams message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns MoveParams
     */
    public static fromObject(object: { [k: string]: any }): MoveParams;

    /**
     * Creates a plain object from a MoveParams message. Also converts values to other types if specified.
     * @param message MoveParams
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: MoveParams, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this MoveParams to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

export namespace MoveParams {

    /** Properties of a NewAnwser. */
    interface INewAnwser {

        /** NewAnwser index */
        index?: (number|null);

        /** NewAnwser gender */
        gender?: (number|null);

        /** NewAnwser emotion */
        emotion?: (number|null);
    }

    /** Represents a NewAnwser. */
    class NewAnwser implements INewAnwser {

        /**
         * Constructs a new NewAnwser.
         * @param [properties] Properties to set
         */
        constructor(properties?: MoveParams.INewAnwser);

        /** NewAnwser index. */
        public index: number;

        /** NewAnwser gender. */
        public gender: number;

        /** NewAnwser emotion. */
        public emotion: number;

        /**
         * Creates a new NewAnwser instance using the specified properties.
         * @param [properties] Properties to set
         * @returns NewAnwser instance
         */
        public static create(properties?: MoveParams.INewAnwser): MoveParams.NewAnwser;

        /**
         * Encodes the specified NewAnwser message. Does not implicitly {@link MoveParams.NewAnwser.verify|verify} messages.
         * @param message NewAnwser message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encode(message: MoveParams.INewAnwser, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Encodes the specified NewAnwser message, length delimited. Does not implicitly {@link MoveParams.NewAnwser.verify|verify} messages.
         * @param message NewAnwser message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encodeDelimited(message: MoveParams.INewAnwser, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Decodes a NewAnwser message from the specified reader or buffer.
         * @param reader Reader or buffer to decode from
         * @param [length] Message length if known beforehand
         * @returns NewAnwser
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): MoveParams.NewAnwser;

        /**
         * Decodes a NewAnwser message from the specified reader or buffer, length delimited.
         * @param reader Reader or buffer to decode from
         * @returns NewAnwser
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): MoveParams.NewAnwser;

        /**
         * Verifies a NewAnwser message.
         * @param message Plain object to verify
         * @returns `null` if valid, otherwise the reason why it is not
         */
        public static verify(message: { [k: string]: any }): (string|null);

        /**
         * Creates a NewAnwser message from a plain object. Also converts values to their respective internal types.
         * @param object Plain object
         * @returns NewAnwser
         */
        public static fromObject(object: { [k: string]: any }): MoveParams.NewAnwser;

        /**
         * Creates a plain object from a NewAnwser message. Also converts values to other types if specified.
         * @param message NewAnwser
         * @param [options] Conversion options
         * @returns Plain object
         */
        public static toObject(message: MoveParams.NewAnwser, options?: $protobuf.IConversionOptions): { [k: string]: any };

        /**
         * Converts this NewAnwser to JSON.
         * @returns JSON object
         */
        public toJSON(): { [k: string]: any };
    }
}

/** Properties of a PushParams. */
export interface IPushParams {
}